Added several new adjustments for the armor part.
added a call to an "AfterApply" method when an adjustment is applied.

added new AfterAdjustmentAppliedEvent and EarlyAfterAdjustmentAppliedEvent MinEvents for immediately before "AfterApply" is called, and at the end.

TODO:
- Add an enumerator for applied adjustments to the Adjustments collection.
- Confirm that adjustments passed to and applied by an operator from a NaturalEquipmentMod have hteir applied status reflected by this.
- finish implementation of DescriptionElement retrieval from Adjustments.

+++ Common/NaturalEquipmentManagement/Adjustments/MeleeWeaponAdjustments/AdjustBonusCap.cs
@@ -34,7 +34,7 @@ namespace HNPS_GigantismPlus
                 MeleeWeapon meleeWeapon = Subject.GetPart<MeleeWeapon>();
                 if (meleeWeapon.MaxStrengthBonus < 999)
                 {
-                    meleeWeapon.AdjustBonusCap(Amount);
+                    meleeWeapon.AdjustBonusCap((int)Amount);
                 }
                 return true;
             }
